Install coc.nvim
As part of this process I've decided to move away from Pathogen for
package management and have started using VimPlug instead. This has the
advantage of allowing me to declare my plugins directly in .vimrc and a
run a single command to install or remove packages.

In order to add Metals support, I've replaced the VimComplete plugin
with coc.nvim (which requires vim 8.1 or nvim - which I refuse to use -
to work). Coc.vim requires later versions of Node to be installed,
meaning that in the `new_machine_setup.sh` file I have had to move the
node installation from the postGNOME setup step and into the vimConfig
step.