Fix issues with the arch script
Have had to install Manjaro on my desktop as Debian 10 is fucking
impossible to use on that as well. Fixing bugs in docker, vim, tmux and
clamAV installations that I noticed this time round. This involved,
installing gvim from the pacman repository (too many errors when
compiling from source this time), ensuring freshclam was run and the
docker service is started among other things.