Make it so you only input the sudo password once
Have added here strings, &c. so that a sudo password will only need to
be input once when the arch script is run. May add something similar for
the Debian version when I have time.
Have also removed the snap installation as it was only used for Spotify
and it hasn't worked on two out of three of the machines I tried to
install it with. There's an AUR package that does the job better.