Copy gitignore properly
The new machine scripts didn't copy the gitignore file properly. Have
fixed these lines in the script.
Have also replaced a hardcoded instance of the home folder in the zshrc
file as the home folders in macOS are obviously in a different location
to the ones in Debian and Manjaro linux.