Improve conky performance
The use of the `top` command and `io.popen` resulted in an unusually
high CPU usage on my machine. Replacing this with the `ps` command run
in `execi` (with a sensible interval value set) seems to have fixed this
issue.
I've also made further improvements to the configuration, in particular
by moving the code that checked the current public IP address into a
separate lua file.